refactor: Split up Commands/Queries into subdomains
