+++ services/peepo/src/application/Command/Alts/UpdateAllAlts.ts
@@ -1,6 +1,6 @@
 import { updateAllAlts } from '@ps2gg/alts/ws'
 import { APIEmbed } from 'discord.js'
-import { getAlts } from '../Query/GetAlts'
+import { getAlts } from '../../Query/Alts/GetAlts'
 
 export async function updateAllAltsCommand(name: string): Promise<APIEmbed> {
   await updateAllAlts(name)
