+++ services/peepo/src/application/Command/Players/VerifyCharacter.ts
@@ -1,5 +1,5 @@
 import { UserClient } from '@ps2gg/users/client'
-import { VerifySuccessEmbed } from '../../domain/Embed/VerifySuccessEmbed'
+import { VerifySuccessEmbed } from '../../../domain/Embed/VerifySuccessEmbed'
 
 export async function verifyCharacter(characterId: string, name: string, discordId: string): Promise<VerifySuccessEmbed> {
   const users = new UserClient()
