+++ services/peepo/src/application/Query/Alts/GetAltsTree.ts
@@ -1,5 +1,5 @@
 import { getAltMatches } from '@ps2gg/alts/ws'
-import { AltTreeEmbed } from '../../domain/Embed/AltTreeEmbed'
+import { AltTreeEmbed } from '../../../domain/Embed/AltTreeEmbed'
 
 export async function getAltsTree(name: string): Promise<AltTreeEmbed> {
   const { result } = await getAltMatches(name)
