+++ services/peepo/src/application/Query/Friends/GetAltWideFriends.ts
@@ -1,11 +1,12 @@
 import { User } from '@ps2gg/users/types'
-import { SeshEmbed } from '../../domain/Embed/SeshEmbed'
-import { getAltIds } from './GetAltIds'
+import { FriendsEmbed } from '../../../domain/Embed/FriendsEmbed'
+import { getAltIds } from '../Alts/GetAltIds'
+import { getOnlinePlayers } from '../Players/GetOnlinePlayers'
+import { getPlayer } from '../Players/GetPlayer'
 import { getFriends } from './GetFriends'
-import { getPlayer } from './GetPlayer'
-import { getSesh } from './GetSesh'
+import { getFriendsWithParent } from './GetFriendsWithParents'
 
-export async function getAltWideSesh(user: User, name: string): Promise<SeshEmbed> {
+export async function getAltWideFriends(user: User, name: string): Promise<FriendsEmbed> {
   let characterIds = user.characterIds?.length ? user.characterIds : []
 
   const player = name ? await getPlayer(name) : null
@@ -20,5 +21,7 @@ export async function getAltWideSesh(user: User, name: string): Promise<SeshEmbe
   const friendsAltIds = friendsAlts.map((alt) => alt.characterId)
   const altWideFriendIds = [...friendIds, ...friendsAltIds]
   const alts = [...playersAlts, ...friendsAlts]
-  return getSesh(player, friendIds, altWideFriendIds, alts)
+  const friends = await getOnlinePlayers(altWideFriendIds)
+  const friendsWithParent = await getFriendsWithParent(friends, friendIds, alts)
+  return new FriendsEmbed(friendsWithParent)
 }
