+++ services/peepo/src/application/Query/Population/GetEventSuggestions.ts
@@ -1,5 +1,5 @@
 import { AutocompleteResponse } from '@ps2gg/discord/command'
-import { PopulationEntity } from '../../domain/Entity/PopulationEntity'
+import { PopulationEntity } from '../../../domain/Entity/PopulationEntity'
 
 export function getEventSuggestions(query: string): AutocompleteResponse[] {
   const events = PopulationEntity.getEvents()
