+++ services/peepo/src/application/Query/Population/GetPopulationNotification.ts
@@ -1,5 +1,5 @@
 import { PopulationClient } from '@ps2gg/population/client'
-import { NotificationEmbed } from '../../domain/Embed/Notification'
+import { NotificationEmbed } from '../../../domain/Embed/Notification'
 
 export async function getPopulationNotification(id: string): Promise<NotificationEmbed> {
   const population = await new PopulationClient().getPopulation(id)
