+++ services/peepo/src/application/Query/Sesh/GetAltWideSesh.ts
@@ -1,12 +1,11 @@
 import { User } from '@ps2gg/users/types'
-import { FriendsEmbed } from '../../domain/Embed/FriendsEmbed'
-import { getAltIds } from './GetAltIds'
-import { getFriends } from './GetFriends'
-import { getFriendsWithParent } from './GetFriendsWithParents'
-import { getOnlinePlayers } from './GetOnlinePlayers'
-import { getPlayer } from './GetPlayer'
+import { SeshEmbed } from '../../../domain/Embed/SeshEmbed'
+import { getAltIds } from '../Alts/GetAltIds'
+import { getFriends } from '../Friends/GetFriends'
+import { getPlayer } from '../Players/GetPlayer'
+import { getSesh } from './GetSesh'
 
-export async function getAltWideFriends(user: User, name: string): Promise<FriendsEmbed> {
+export async function getAltWideSesh(user: User, name: string): Promise<SeshEmbed> {
   let characterIds = user.characterIds?.length ? user.characterIds : []
 
   const player = name ? await getPlayer(name) : null
@@ -21,7 +20,5 @@ export async function getAltWideFriends(user: User, name: string): Promise<Frien
   const friendsAltIds = friendsAlts.map((alt) => alt.characterId)
   const altWideFriendIds = [...friendIds, ...friendsAltIds]
   const alts = [...playersAlts, ...friendsAlts]
-  const friends = await getOnlinePlayers(altWideFriendIds)
-  const friendsWithParent = await getFriendsWithParent(friends, friendIds, alts)
-  return new FriendsEmbed(friendsWithParent)
+  return getSesh(player, friendIds, altWideFriendIds, alts)
 }
