+++ services/peepo/src/application/Query/Sesh/GetSesh.ts
@@ -1,8 +1,8 @@
 import { Player } from '@ps2gg/players/types'
-import { SeshEmbed } from '../../domain/Embed/SeshEmbed'
-import { AltMatch } from './GetAltIds'
-import { getFriendsWithParent } from './GetFriendsWithParents'
-import { getOnlinePlayers } from './GetOnlinePlayers'
+import { SeshEmbed } from '../../../domain/Embed/SeshEmbed'
+import { AltMatch } from '../Alts/GetAltIds'
+import { getFriendsWithParent } from '../Friends/GetFriendsWithParents'
+import { getOnlinePlayers } from '../Players/GetOnlinePlayers'
 
 export async function getSesh(player: Player, friendIds: string[], altWideFriendIds: string[], alts: AltMatch[]): Promise<SeshEmbed> {
   const friends = await getOnlinePlayers(altWideFriendIds)
