+++ services/peepo/src/infrastructure/SlashCommand/Alt.ts
@@ -2,9 +2,9 @@ import { getPlayerAutocomplete } from '@ps2gg/alts/ws'
 import { sanitizeCharacterName } from '@ps2gg/common/util'
 import { Autocomplete, AutocompleteResponse, Command, Component, ComponentResponse, SubCommand, CommandResponse } from '@ps2gg/discord/command'
 import { CommandInteraction, MessageComponentInteraction } from 'discord.js'
-import { updateAllAltsCommand } from '../../application/Command/UpdateAllAlts'
-import { getAlts } from '../../application/Query/GetAlts'
-import { getAltsTree } from '../../application/Query/GetAltsTree'
+import { updateAllAltsCommand } from '../../application/Command/Alts/UpdateAllAlts'
+import { getAlts } from '../../application/Query/Alts/GetAlts'
+import { getAltsTree } from '../../application/Query/Alts/GetAltsTree'
 import { AltExplain } from '../../domain/Component/AltExplain'
 import { AltReset } from '../../domain/Component/AltReset'
 import { AltUpdate } from '../../domain/Component/AltUpdate'
