+++ services/peepo/src/infrastructure/SlashCommand/Friends.ts
@@ -2,8 +2,8 @@ import { getPlayerAutocomplete } from '@ps2gg/alts/ws'
 import { Autocomplete, AutocompleteResponse, Command, CommandResponse, Main, linkedUser } from '@ps2gg/discord/command'
 import { User } from '@ps2gg/users/types'
 import { CommandInteraction } from 'discord.js'
-import { validateVerification } from '../../application/Command/ValidateVerification'
-import { getAltWideFriends } from '../../application/Query/GetAltWideFriends'
+import { validateVerification } from '../../application/Command/Users/ValidateVerification'
+import { getAltWideFriends } from '../../application/Query/Friends/GetAltWideFriends'
 import { VerifyHintEmbed } from '../../domain/Embed/VerifyHintEmbed'
 import { Friends, FriendsOptions } from '../../domain/Meta/Friends'
