+++ services/peepo/src/infrastructure/SlashCommand/Notify.ts
@@ -2,13 +2,13 @@ import { sanitizeObjectNotation } from '@ps2gg/common/util'
 import { Component, Command, linkedUser, Autocomplete, AutocompleteResponse, CommandResponse, Main, ComponentResponse, ProgressiveComponent } from '@ps2gg/discord/command'
 import { User } from '@ps2gg/users/types'
 import { AnySelectMenuInteraction, ButtonInteraction, CommandInteraction } from 'discord.js'
-import { addPopulationSubscription } from '../../application/Command/AddPopulationSubscription'
-import { modifyPopulationSubscriptionConfig } from '../../application/Command/ModifyPopulationSubscriptionConfig'
-import { modifyPopulationSubscriptionTime } from '../../application/Command/ModifyPopulationSubscriptionTime'
-import { removePopulationSubscription } from '../../application/Command/RemovePopulationSubscription'
-import { getEventSuggestions } from '../../application/Query/GetEventSuggestions'
-import { getPopulationSubscriptionIds } from '../../application/Query/GetPopulationSubscriptionIds'
-import { getSubscription } from '../../application/Query/GetSubscription'
+import { addPopulationSubscription } from '../../application/Command/Population/AddPopulationSubscription'
+import { modifyPopulationSubscriptionConfig } from '../../application/Command/Population/ModifyPopulationSubscriptionConfig'
+import { modifyPopulationSubscriptionTime } from '../../application/Command/Population/ModifyPopulationSubscriptionTime'
+import { removePopulationSubscription } from '../../application/Command/Population/RemovePopulationSubscription'
+import { getEventSuggestions } from '../../application/Query/Population/GetEventSuggestions'
+import { getPopulationSubscriptionIds } from '../../application/Query/Population/GetPopulationSubscriptionIds'
+import { getSubscription } from '../../application/Query/Population/GetSubscription'
 import { NotifyConfigurePopulation } from '../../domain/Component/NotifyConfigurePopulation'
 import { NotifyConfigureSendAfter, NotifyConfigureSendBefore } from '../../domain/Component/NotifyConfigureTime'
 import { Unsubscribe } from '../../domain/Component/Unsubscribe'
