+++ services/peepo/src/infrastructure/SlashCommand/Verify.ts
@@ -3,8 +3,8 @@ import { sanitizeCharacterName } from '@ps2gg/common/util'
 import { Command, Main, Autocomplete, AutocompleteResponse, CommandResponse, Component, ComponentResponse } from '@ps2gg/discord/command'
 import { DiscordCommand } from '@ps2gg/discord/types'
 import { ButtonInteraction } from 'discord.js'
-import { verifyCharacter } from '../../application/Command/VerifyCharacter'
-import { getSelectedCharacter } from '../../application/Query/GetSelectedCharacter'
+import { verifyCharacter } from '../../application/Command/Players/VerifyCharacter'
+import { getSelectedCharacter } from '../../application/Query/Players/GetSelectedCharacter'
 import { VerifyReady } from '../../domain/Component/VerifyReady'
 import { verificationCoordinator } from '../../domain/Coordinator/VerificationCoordinator'
 import { Verify, VerifyOptions } from '../../domain/Meta/Verify'
