+++ services/peepo/src/infrastructure/SlashCommand/WhoAmI.ts
@@ -1,7 +1,7 @@
 import { Command, Main, CommandResponse, linkedUser } from '@ps2gg/discord/command'
 import { DiscordCommand } from '@ps2gg/discord/types'
 import { User } from '@ps2gg/users/types'
-import { getPlayers } from '../../application/Query/GetPlayers'
+import { getPlayers } from '../../application/Query/Players/GetPlayers'
 import { VerifiedCharactersEmbed } from '../../domain/Embed/VerifiedCharactersEmbed'
 import { WhoAmI } from '../../domain/Meta/WhoAmI'
